Refactor WorkListTest.
