+++ src/main/java/seedu/address/model/EquipmentManager.java
@@ -11,7 +11,6 @@ import javafx.collections.ObservableList;
 import seedu.address.commons.util.InvalidationListenerManager;
 import seedu.address.model.equipment.Equipment;
 import seedu.address.model.equipment.UniqueEquipmentList;
-import seedu.address.model.equipment.UniqueWorkListList;
 import seedu.address.model.tag.Tag;
 
 /**
