+++ src/main/java/seedu/address/model/UniqueWorkListList.java
@@ -1,4 +1,4 @@
-package seedu.address.model.equipment;
+package seedu.address.model;
 
 import static java.util.Objects.requireNonNull;
 //import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
@@ -9,7 +9,6 @@ import java.util.List;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import seedu.address.model.WorkList;
 import seedu.address.model.equipment.exceptions.DuplicateEquipmentException;
 import seedu.address.model.equipment.exceptions.EquipmentNotFoundException;
