+++ src/test/java/seedu/address/model/UniqueWorkListListTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.equipment;
+package seedu.address.model;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -21,6 +21,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
+import seedu.address.model.UniqueWorkListList;
 import seedu.address.model.WorkList;
 import seedu.address.model.equipment.exceptions.DuplicateEquipmentException;
 import seedu.address.model.equipment.exceptions.EquipmentNotFoundException;
